fixed line number issue and load issues
