--- tests/evaluation/load/A.csv
@@ -1,1 +1,0 @@
-1	2	3
