--- tests/evaluation/load/B.csv
@@ -1,2 +1,0 @@
-1	2	3
-1	45	n/a
