--- tests/evaluation/load/C.csv
@@ -1,2 +1,0 @@
-1	2	3
-1	2	4
