+++ tests/evaluation/load/load.err
@@ -1,2 +1,1 @@
 Wrong arity of fact file A.facts!
-Wrong arity of fact file B.facts!
